backend/requirements.txt
flask
flask_cors
flask-jwt-extended
pymongo
datatime
urllib.parse
platform
socket
re
secrets
bson
werkzeug.security